Make DWORD exactly 32 bits, so 64-bit lays out data like 32-bit does
os_unix.h spelled the base types as "unsigned long" and annotated DWORD with
"// 32 bits". That is true on a 32-bit build and wrong on a 64-bit one, where
unsigned long is 8 bytes. Everything built on DWORD doubled in width there.

Measured over the 47 named members of the packet unions in grayproto.h, 29 had
a different size on the two builds - Party_Msg_Data was 14 bytes on 32-bit and
26 on 64-bit. That is the packet corruption the README warns about. After this
change every one of the 47 has the same offset and size on both, and
grayproto.h carries static_asserts on the wire type widths so it stays that way.

The types are now uint8_t / uint16_t / uint32_t, LONGLONG is int64_t (it was
"unsigned long", so it was only 32 bits on the reference build), and LONG is
int32_t.

Three things had to follow:

- twofish/aes.h redefined BYTE and DWORD as "unsigned long" for the rest of
  every translation unit that reached it. Twofish is defined over 32-bit words,
  so on 64-bit the key schedule, the S-boxes and the subkeys were all twice the
  intended width. It now takes the fixed width types, or whatever the platform
  already defined.

- MAKEINTRESOURCE hides a small id inside a pointer and ISINTRESOURCE tests the
  high bits to tell an id from a real string. Both went through DWORD, which
  once DWORD is 32 bits truncates a 64-bit pointer - so a string pointer whose
  low half looked small would be read back as an id. They go through uintptr_t
  now, which is the same 32-bit mask on the reference build.

- 367 format specifiers printed a DWORD with %lx / %lu / %lX. The output text is
  byte identical for the whole 32-bit range, verified on both architectures, so
  no save file or log line changes; the ones that really did take a long are
  left alone. NetState's id was a long only to match those, and is an index into
  the client array, so it is an int.

Two uses of ULONG_MAX as an all-bits DWORD became UINT_MAX: on 64-bit they used
to produce ffffffffffffffff, which the script side saw as a different value than
on 32-bit.

+++ src/graysvr/CWebPage.cpp
@@ -623,7 +623,7 @@ int CWebPageDef::ServPageRequest( CClient * pClient, LPCTSTR pszURLArgs, CGTime 
 		iLen += sprintf(szTmp + iLen, "Last-Modified: %s\r\n",  CGTime(dateChange).FormatGmt(NULL));
 
 	iLen += sprintf( szTmp + iLen,
-		"Content-Length: %lu\r\n"
+		"Content-Length: %u\r\n"
 		"\r\n",
 		dwSize
 		);
@@ -719,7 +719,7 @@ bool CWebPageDef::ServPagePost( CClient * pClient, LPCTSTR pszURLArgs, TCHAR * p
 	// C or CHK or CHECK = the check boxes
 
 	CDialogResponseArgs resp;
-	DWORD dwButtonID = ULONG_MAX;
+	DWORD dwButtonID = UINT_MAX;	// DWORD is 32 bits, ULONG_MAX is not
 	for ( size_t i = 0; i < iArgs; i++ )
 	{
 		TCHAR * pszNum = ppArgs[i];
